+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## [unreleased]
 ### Changes
 - Replace QUAST with MetaQUAST
 - Add species ID to run summary table
+- Tidy up presentation in report
 - `-profile conda` is no longer supported, users should use `-profile standard` (Docker) or `-profile singularity` instead
 ### Added
 - `nextflow run epi2me-labs/wf-human-variation --version` will now print the workflow version number and exit